Sync

+++ include/utils.h
@@ -26,7 +26,7 @@
 #define SDU_ARP_TYPE_MATCH  1
 
 extern int debug_mode;
-extern int localMIP;
+extern uint8_t localMIP;
